Moves Carousel => FlickityCarousel
