+++ src/v2/Apps/Artist/Components/ArtistTopWorksRail/ArtistTopWorksRail.tsx
@@ -2,7 +2,7 @@ import { Box, Flex, Sans } from "@artsy/palette"
 import React from "react"
 import { createFragmentContainer, graphql } from "react-relay"
 import { ArtistTopWorksRail_artist } from "v2/__generated__/ArtistTopWorksRail_artist.graphql"
-import { Carousel } from "v2/Components/Carousel"
+import { Carousel } from "v2/Components/FlickityCarousel"
 import FillwidthItem from "v2/Components/Artwork/FillwidthItem"
 import { Media } from "v2/Utils/Responsive"
 import { StyledLink } from "v2/Apps/Artist/Components/StyledLink"
