+++ src/v2/Apps/Artist/Components/ArtistTopWorksRail/__tests__/ArtistTopWorksRail.jest.tsx
@@ -2,7 +2,7 @@ import { mount } from "enzyme"
 import React from "react"
 import { ArtistTopWorksRailFragmentContainer as ArtistTopWorksRail } from "../ArtistTopWorksRail"
 import { ArtistTopWorksRail_artist } from "v2/__generated__/ArtistTopWorksRail_artist.graphql"
-import { Carousel } from "v2/Components/Carousel"
+import { Carousel } from "v2/Components/FlickityCarousel"
 import FillwidthItem from "v2/Components/Artwork/FillwidthItem"
 import { useTracking } from "v2/Artsy/Analytics/useTracking"
 import { StyledLink } from "v2/Apps/Artist/Components/StyledLink"
