+++ src/v2/Apps/Artist/Routes/Overview/Components/RecommendedArtist.tsx
@@ -5,7 +5,7 @@ import { SystemContext } from "v2/Artsy"
 import { track } from "v2/Artsy/Analytics"
 import * as Schema from "v2/Artsy/Analytics/Schema"
 import FillwidthItem from "v2/Components/Artwork/FillwidthItem"
-import { ArrowButton, Carousel } from "v2/Components/Carousel"
+import { ArrowButton, Carousel } from "v2/Components/FlickityCarousel"
 import { FollowArtistButtonFragmentContainer as FollowArtistButton } from "v2/Components/FollowButton/FollowArtistButton"
 import React, { FC, useContext } from "react"
 import { createFragmentContainer, graphql } from "react-relay"
