+++ src/v2/Apps/Artist/Routes/Overview/Components/WorksForSaleRail.tsx
@@ -8,7 +8,7 @@ import * as Schema from "v2/Artsy/Analytics/Schema"
 import { renderWithLoadProgress } from "v2/Artsy/Relay/renderWithLoadProgress"
 import { SystemQueryRenderer } from "v2/Artsy/Relay/SystemQueryRenderer"
 import FillwidthItem from "v2/Components/Artwork/FillwidthItem"
-import { ArrowButton, Carousel } from "v2/Components/Carousel"
+import { ArrowButton, Carousel } from "v2/Components/FlickityCarousel"
 import React, { useContext } from "react"
 import { createFragmentContainer, graphql } from "react-relay"
 import styled from "styled-components"
