+++ src/v2/Apps/Artwork/Components/ArtworkArtistSeries/ArtistSeriesArtworkRail.tsx
@@ -2,7 +2,7 @@ import { Box, Flex, Sans } from "@artsy/palette"
 import React from "react"
 import { createFragmentContainer, graphql } from "react-relay"
 import { ArtistSeriesArtworkRail_artwork } from "v2/__generated__/ArtistSeriesArtworkRail_artwork.graphql"
-import { Carousel } from "v2/Components/Carousel"
+import { Carousel } from "v2/Components/FlickityCarousel"
 import FillwidthItem from "v2/Components/Artwork/FillwidthItem"
 import { StyledLink } from "v2/Apps/Artist/Components/StyledLink"
 import { getENV } from "v2/Utils/getENV"
