+++ src/v2/Apps/Collect/Routes/Collection/Components/CollectionsHubRails/ArtistSeriesRail/__tests__/ArtistSeriesRail.jest.tsx
@@ -1,6 +1,6 @@
 import { CollectionsHubLinkedCollections } from "v2/Apps/__tests__/Fixtures/Collections"
 import { useTracking } from "v2/Artsy/Analytics/useTracking"
-import { ArrowButton } from "v2/Components/Carousel"
+import { ArrowButton } from "v2/Components/FlickityCarousel"
 import { mount } from "enzyme"
 import "jest-styled-components"
 import React from "react"
@@ -68,10 +68,7 @@ describe("ArtistSeriesRail", () => {
       ]
 
       const component = mount(<ArtistSeriesRail {...props} />)
-      component
-        .find(ArrowButton)
-        .at(1)
-        .simulate("click")
+      component.find(ArrowButton).at(1).simulate("click")
 
       expect(trackEvent).toBeCalledWith({
         action_type: "Click",
