+++ src/v2/Apps/Collect/Routes/Collection/Components/CollectionsHubRails/FeaturedCollectionsRails/__tests__/FeaturedCollectionsRails.jest.tsx
@@ -1,6 +1,6 @@
 import { CollectionHubFixture } from "v2/Apps/__tests__/Fixtures/Collections"
 import { useTracking } from "v2/Artsy/Analytics/useTracking"
-import { ArrowButton } from "v2/Components/Carousel"
+import { ArrowButton } from "v2/Components/FlickityCarousel"
 import { mount } from "enzyme"
 import "jest-styled-components"
 import React from "react"
@@ -65,10 +65,7 @@ describe("FeaturedCollectionsRails", () => {
 
       const component = mount(<FeaturedCollectionsRails {...props} />)
 
-      component
-        .find(ArrowButton)
-        .at(1)
-        .simulate("click")
+      component.find(ArrowButton).at(1).simulate("click")
 
       expect(trackEvent).toBeCalledWith({
         action_type: "Click",
@@ -133,10 +130,7 @@ describe("FeaturedCollectionEntity", () => {
     const component = mount(
       <FeaturedCollectionEntity member={members[0]} itemNumber={0} />
     )
-    component
-      .find(StyledLink)
-      .at(0)
-      .simulate("click")
+    component.find(StyledLink).at(0).simulate("click")
 
     expect(trackEvent).toBeCalledWith({
       action_type: "Click",
