+++ src/v2/Components/ArtistSeriesRail/ArtistSeriesRail.tsx
@@ -1,5 +1,5 @@
 import { Box, Sans } from "@artsy/palette"
-import { ArrowButton, Carousel } from "v2/Components/Carousel"
+import { ArrowButton, Carousel } from "v2/Components/FlickityCarousel"
 import React from "react"
 import { createFragmentContainer, graphql } from "react-relay"
 import styled from "styled-components"
